car_dealer_app/requirements.txt
Flask==2.2.5
gunicorn==20.1.0
google-cloud-datastore
google-cloud-storage
